Create a test harness for testing split/aggregation utils
